Built-In-Tools-Starter-Example/requirements.txt
langchain
langchain-community
langchain-experimental
boto3
python-dotenv
wikipedia
duckduckgo-search 